refactor: move actions
