+++ packages/camphora-styled/src/button.css.ts
@@ -1,6 +1,6 @@
 import { style } from "@vanilla-extract/css";
 
-import { vars } from "../theme.css";
+import { vars } from "./theme.css";
 
 export const buttonActive = style({});
