+++ packages/camphora-styled/src/index.ts
@@ -1,8 +1,6 @@
 export * from "./actions";
-export * from "./actions/button.css";
-export * from "./actions/popover.css";
+export * from "./button.css";
 export * from "./card.css";
-export * from "./drawer.css";
 export * from "./navbar.css";
 export * from "./previewBlock.css";
 export * from "./theme.css";
